+++ cmd/vega/genesis/update.go
@@ -8,10 +8,8 @@ import (
 
 	vgfs "code.vegaprotocol.io/shared/libs/fs"
 	"code.vegaprotocol.io/shared/paths"
-	"code.vegaprotocol.io/vega/assets"
 	"code.vegaprotocol.io/vega/genesis"
 	"code.vegaprotocol.io/vega/logging"
-	"code.vegaprotocol.io/vega/netparams"
 	"code.vegaprotocol.io/vega/nodewallets"
 	"code.vegaprotocol.io/vega/validators"
 
@@ -24,9 +22,8 @@ import (
 type updateCmd struct {
 	Config nodewallets.Config
 
-	DryRun  bool   `long:"dry-run" description:"Display the genesis file without writing it"`
-	Network string `short:"n" long:"network" choice:"mainnet" choice:"testnet"`
-	TmHome  string `short:"t" long:"tm-home" description:"The home path of tendermint"`
+	DryRun bool   `long:"dry-run" description:"Display the genesis file without writing it"`
+	TmHome string `short:"t" long:"tm-home" description:"The home path of tendermint"`
 }
 
 func (opts *updateCmd) Execute(_ []string) error {
@@ -69,28 +66,9 @@ func (opts *updateCmd) Execute(_ []string) error {
 		TmPubKey:        b64TmPubKey,
 	}
 
-	if len(opts.Network) != 0 {
-		ethConfig := `{"network_id": "%s", "chain_id": "%s", "bridge_address": "%s", "confirmations": %d,  "staking_bridge_addresses": %s}`
-		switch opts.Network {
-		case "mainnet":
-			delete(genesisState.Assets, "VOTE")
-			genesisState.Assets["VEGA"] = assets.VegaTokenMainNet
-			marshalledBridgeAddresses, _ := json.Marshal([]string{"0xfc9Ad8fE9E0b168999Ee7547797BC39D55d607AA", "0x1B57E5393d949242a9AD6E029E2f8A684BFbBC08"})
-			ethConfig = fmt.Sprintf(ethConfig, "3", "3", "0x898b9F9f9Cab971d9Ceb809F93799109Abbe2D10", 3, marshalledBridgeAddresses)
-		case "testnet":
-			genesisState.Assets["VEGA"] = assets.VegaTokenTestNet
-			delete(genesisState.Assets, "VOTE")
-			marshalledBridgeAddresses, _ := json.Marshal([]string{"0xfc9Ad8fE9E0b168999Ee7547797BC39D55d607AA", "0x1B57E5393d949242a9AD6E029E2f8A684BFbBC08"})
-			ethConfig = fmt.Sprintf(ethConfig, "3", "3", "0x898b9F9f9Cab971d9Ceb809F93799109Abbe2D10", 3, marshalledBridgeAddresses)
-		default:
-			return fmt.Errorf("network %s is not supported", opts.Network)
-		}
-		genesisState.NetParams[netparams.BlockchainsEthereumConfig] = ethConfig
-	}
-
 	rawGenesisState, err := json.Marshal(genesisState)
 	if err != nil {
-		return err
+		return fmt.Errorf("couldn't marshal the genesis state as JSON: %w", err)
 	}
 
 	genesisFilePath := tmConfig.GenesisFile()
@@ -102,7 +80,7 @@ func (opts *updateCmd) Execute(_ []string) error {
 	genesisDoc := &types.GenesisDoc{}
 	err = tmjson.Unmarshal(data, genesisDoc)
 	if err != nil {
-		return err
+		return fmt.Errorf("couldn't unmarshal the genesis document: %w", err)
 	}
 
 	genesisDoc.AppState = rawGenesisState
@@ -110,13 +88,13 @@ func (opts *updateCmd) Execute(_ []string) error {
 	if !opts.DryRun {
 		log.Infof("Saving genesis doc at %s", genesisFilePath)
 		if err := genesisDoc.SaveAs(genesisFilePath); err != nil {
-			return err
+			return fmt.Errorf("couldn't save the genesis file: %w", err)
 		}
 	}
 
 	marshalledGenesisDoc, err := tmjson.MarshalIndent(genesisDoc, "", "  ")
 	if err != nil {
-		return err
+		return fmt.Errorf("couldn't marshal the genesis document as JSON: %w", err)
 	}
 	fmt.Println(string(marshalledGenesisDoc))
 	return err