+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [4581] (https://github.com/vegaprotocol/vega/issues/4561) - Separate endpoints for liquidity provision submissions, amendment and cancellation
 - [4390](https://github.com/vegaprotocol/vega/pull/4390) - Introduce node mode, `vega init` now require a mode: full or validator
 - [4383](https://github.com/vegaprotocol/vega/pull/4383) - Rename flag `--tm-root` to `--tm-home`
+- [](https://github.com/vegaprotocol/vega/pull/) - Remove the outdated `--network` flags on `vega genesis generate` and `vega genesis update`
 
 ### 🗑️ Deprecation
 - [](https://github.com/vegaprotocol/vega/pull/) -